[IMP] Make log work on request level and make it prettier
